fixed controller and service layers in accordance with new DTO models

+++ apps/visualizer/static/visualizer/vue/dataset-editor.js
@@ -5,11 +5,13 @@ export default {
     components: {
         'dataset-editor-datarow': () => import(staticFiles + "vue/dataset-editor-datarow.js"),
     },
-    props: ['result'],
+    props: ['result', 'new_dataset'],
     data() {
         return {
             filename: this.result.name,
-            file_id: this.result.file_id,
+            file_id: null,
+            id: null,
+            timestamp: this.result.timestamp,
             width: this.result.width,
             height: this.result.height,
             column_names: this.result.column_names,
@@ -21,6 +23,13 @@ export default {
     },
 
     created: function () {
+        if (this.new_dataset) {
+            this.file_id = this.result.file_id
+        } else {
+            this.id = this.result.id
+        }
+        this.file_id = this.result.file_id,
+        this.id = this.result.id,
         window.addEventListener("beforeunload", this.preventNav);
         this.$once("hook:beforeDestroy", () => {
           window.removeEventListener("beforeunload", this.preventNav);
@@ -46,9 +55,9 @@ export default {
             this.column_types[index] = event.target.value;
         },
 
-        onSave: function() {
+        onCreate: function() {
             this.edit = true;
-            fetch('/api/save', {
+            fetch('/api/create', {
                 method: 'POST',
                 body: JSON.stringify({
                     name: this.filename,
@@ -67,19 +76,48 @@ export default {
                 });
             });
         },
-        onCancel: function() {
+
+        onSave: function() {
             this.edit = true;
-            fetch(`/api/editor/${this.file_id}`, {
-                method: 'DELETE',
+            fetch(`/api/datasets/${this.id}`, {
+                method: 'PUT',
+                body: JSON.stringify({
+                    name: this.filename,
+                    id: this.id,
+                    timestamp: this.timestamp,
+                    width: this.width,
+                    height: this.height,
+                    column_names: this.column_names,
+                    column_types: this.column_types,
+                    datarows: this.rows
+                })
             })
             .then(response => response.json())
             .then(result => {
                 router.push({
                     name: 'home',
                 });
-            })
+            });
+        },
 
+        onCancel: function() {
+            this.edit = true;
+            if (this.new_dataset) {
+                fetch(`/api/create/${this.file_id}`, {
+                    method: 'DELETE',
+                })
+                .then(response => response.json())
+                .then(result => {
+                    router.push({
+                        name: 'home',
+                    });
+                });
+            } else {
+                router.push({
+                    name: 'home',
+                });
+            }
         }
-    },
 
+    },
 }
